requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
requests>=2.31.0
pyinstaller>=5.13.0